Bug 19893: Alternative optimized indexing for Elasticsearch
Implement optimized indexing for Elasticsearch

How to test:
1) Time a full elasticsearch re-index without this patch by running the
   rebuild_elastic_search.pl with the -d flag:
   `koha-shell <instance_name> -c "time rebuild_elastic_search.pl -d"`.
2) Apply this patch.
3) Time a full re-index again, it should be about twice at fast (for a
   couple of thousand biblios, with fewer biblios results may be more
   unpredictable).

Sponsored-by: Gothenburg University Library

+++ misc/search_tools/rebuild_elastic_search.pl
@@ -160,8 +160,13 @@ sub do_reindex {
     my ( $next, $index_name ) = @_;
 
     my $indexer = Koha::SearchEngine::Elasticsearch::Indexer->new( { index => $index_name } );
+
     if ($delete) {
-        $indexer->drop_index();
+        $indexer->drop_index() if $indexer->index_exists();
+        $indexer->create_index();
+    }
+    elsif (!$indexer->index_exists()) {
+        # Create index if does not exist
         $indexer->create_index();
     }
